worked closures, transitions are wip

+++ parselib/parsers/naiveparsers.hpp
@@ -66,12 +66,12 @@ private :
 
 	void make_list() ;
 
-	void make_regex(int j) ;
+	void make_regex(size_t j) ;
 		
 	/// add to keeper to tell the parser to save this node's content
-	void add_to_keeper(int j) ;
+	void add_to_keeper(size_t j) ;
 
-	void add_to_str_rules(int j) ;
+	void add_to_str_rules(size_t j) ;
 
 } ;
